Graphics: Move Gal/Texture to Texture
